requirements.txt
pydantic>=2.11.0
pytest>=8.4.0
httpx>=0.28.0
python-dotenv>=1.0.1
google-genai>=1.17.0
google-cloud-storage>=2.19.0
fastapi>=0.116.0
uvicorn>=0.35.0
yt-dlp>=2025.3.31
